MueLu: start to migrate tutorial tests to Tpetra
This starts to convert a Laplace2D test from Epetra to Tpetra. We still
stick to the Xpetra world for compatibility with MueLu's ETI system for
unit tests.

So far, the test can setup a Laplace2D system and solve it with Amesos2
(to have an exact solution for comparisons).

Part of #3758.