Update ctest-driver scripts to use different ctest_coverage function
Switch to using `ctest_coverage` instead of `ctest_coverage_collect_gcov` for the
collection of coverage data.

For some time now, we've had issues with the uploading of the generated tar'd gcov files
from `ctest_coverage_collect_gcov`. After experimenting with the simplier `ctest_coverage` function,
we've found this to collect coverage results MUCH faster into a series of `Coverage.xml` that were much
easier to upload through the use of `ctest_submit(PARTS Coverage)` to CDash.

Additonal changes include using our `submit_by_parts` wrapper for the coverage step which is
also being used for the configure, build, and test steps.

The downside of using `submit_by_parts` is that
for the coverage stage, there are two sub-stages that are possible for failures, coverage collection
and coverage upload. If either of these fail, the coverage stage should return a failure result to the
ctest-driver. Since `submit_by_parts` is its own general macro for `ctest_submit`, this finer grain of
control propogating an error between the two substages is lost.

The upside of using `submit_by_parts` is the obviously less repetition of codes related to the
submission step. This switch to `submit_by_parts` was done with the idea of later implementing
multiple CDash submits which is currently not supported.

Signed-off-by: Anderson Chauphan <achauph@sandia.gov>
Signed-off-by: Christian Glusa <caglusa@sandia.gov>